+++ daemon_other.go
@@ -1,0 +1,11 @@
+//go:build !windows
+
+package main
+
+import "fmt"
+
+func hideConsoleWindow() {}
+
+func stopDaemon() {
+	fmt.Println("stop 命令仅在 Windows 平台可用")
+}